requirements.txt
absl-py==0.9.0
aiohttp==3.7.3
appdirs==1.4.4
astor==0.8.1
async-timeout==3.0.1
attrs==19.3.0
audioread==2.1.8
auraloss==0.1.7
backcall==0.1.0
bleach==3.1.0
cachetools==4.2.0
certifi==2019.11.28
cffi==1.13.2
chardet==3.0.4
click==7.1.2
configparser==5.0.1
cycler==0.10.0
dataclasses==0.8
decorator==4.4.1
defusedxml==0.6.0
docker-pycreds==0.4.0
-e git+https://github.com/rastegah/stmetric.git@4baa1cd39fc557471ba6cbe67a2cf25a7e9f3f7a#egg=stmetric
efficientnet-pytorch==0.7.1
entrypoints==0.3
fire==0.3.1
fsspec==2021.6.1
future==0.18.2
gast==0.3.3
gin-config==0.4.0
gitdb==4.0.5
GitPython==3.1.13
google-auth==1.24.0
google-auth-oauthlib==0.4.2
google-pasta==0.1.8
grpcio==1.26.0
h5py==2.10.0
idna==2.8
idna-ssl==1.1.0
importlib-metadata==1.4.0
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.0
joblib==0.14.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterthemes==0.20.0
Keras==2.2.4
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kymatio==0.2.0
lesscpy==0.14.0
librosa==0.7.2
llvmlite==0.31.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.2
mistune==0.8.4
mock==3.0.5
multidict==5.1.0
nbconvert==5.6.1
nbformat==5.0.4
nltk==3.4.5
nnAudio==0.2.1
notebook==6.0.3
numba==0.47.0
numpy==1.18.1
oauthlib==3.1.0
packaging==20.9
pandas==0.25.3
pandocfilters==1.4.2
parso==0.6.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.1.0
ply==3.11
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.3
protobuf==3.15.1
psutil==5.8.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pyDeprecate==0.3.0
pydub==0.23.1
Pygments==2.5.2
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
pytorch-lightning==1.3.8
pytorch-pcen==0.0.1
pytz==2019.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
requests==2.22.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.7
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
Send2Trash==1.5.0
sentry-sdk==0.20.3
shortuuid==1.0.1
six==1.14.0
smmap==3.0.5
SoundFile==0.10.3.post1
subprocess32==3.5.4
tensorboard==2.4.1
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
tensorflow-hub==0.7.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tf-slim==1.0
-e git+https://github.com/rastegah/timbre-augmentations.git@adb854a312d21826f7e92f6344d2c08290da84e9#egg=timbre_augmentations
torch==1.7.1
torchaudio==0.7.2
torchcrepe==0.0.12
torchmetrics==0.4.1
torchvision==0.8.2
tornado==6.0.3
tqdm==4.42.0
traitlets==4.3.3
typing-extensions==3.7.4.3
urllib3==1.25.8
wandb==0.10.20
wavaugment==0.2
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.11.2
yarl==1.6.3
zipp==2.1.0
